// src/models/Food.ts
interface Food {
  _id: string;
  id: string;
  id_v4: string;
  fir_id: string;
  name: string;
  synonyms?: any;
  category: string;
  density?: any;
  unit: string;
  kj: number;
  kcal: number;
  fat_g: number;
  fat_saturated_g: number;
  fat_monounsaturated_g: number;
  fat_polyunsaturated_g: number;
  cholesterol_mg: number;
  carbohydrates_g: number;
  sugars_g: number;
  starch_g?: any;
  dietary_fibres_g: number;
  protein_g: number;
  salt_nacl_g: number;
  alcohol_g: number;
  water_g: number;
  vitamin_a_re_ug: number;
  vitamin_a_rae_ug: number;
  retinol_ug: number;
  betacarotene_bce_ug: number;
  betacarotene_ug: number;
  vitamin_b1_thiamine_mg: number;
  vitamin_b2_riboflavin_mg: number;
  vitamin_b6_pyridoxine_mg: number;
  vitamin_b12_cobalamin_ug: number;
  niacin_mg: number;
  folate_ug: number;
  panthotenic_acid_mg: number;
  vitamin_c_ascorbic_acid_mg: number;
  vitamin_d_calciferol_ug: number;
  vitamin_e_mg: number;
  potassium_k_mg: number;
  sodium_na_mg: number;
  chloride_cl_mg?: any;
  calcium_ca_mg: number;
  magnesium_mg_mg: number;
  phosphorus_p_mg: number;
  iron_fe_mg: number;
  iodide_i_ug: number;
  zinc_zn_mg: number;
  selenium_se_ug?: any;
}
export default Food;